Moved ensure_nearby function to server/src/sim.rs
